refactor: 구조 정일 util
